Put all traits into single module
