+++ Cargo.toml
@@ -16,6 +16,4 @@ name = "blackjack"
 
 [dependencies]
 num = "^0.2"
-num-traits = "^0.2"
 ndarray = "^0.11"
-libc = "^0.2"
