+++ src/prelude.rs
@@ -1,8 +1,6 @@
 //! Default and recommended imports for functionality of crate.
 
-pub use traits::BlackJackData;
+pub use traits::*;
+pub use enums::*;
 pub use series::{Series};
-pub use series::traits::SeriesTrait;
-pub use enums::{SeriesEnumRef, DType};
 pub use dataframe::{DataFrame};
-pub use dataframe::traits::{ColumnManager};
