+++ tests/dataframe_tests.rs
@@ -38,9 +38,9 @@ fn test_add_columns_unknown_types() {
 
     df.add_column(series);
 
-    let seriesenum: SeriesEnumRef = df.get_column_unknown_type("series-1").expect("No column named 'series-1'");
+    let seriesenum: SeriesEnum = df.get_column_unknown_type("series-1").expect("No column named 'series-1'");
 
-    if let SeriesEnumRef::I32(series) = seriesenum {
+    if let SeriesEnum::I32(series) = seriesenum {
         assert_eq!(series.sum(), 10);
     } else {
         panic!("Expected series of I32 but got something else!");
